Tidying a few variables.

+++ tools/orac_main.py
@@ -94,7 +94,8 @@ def orac_main(args):
                 print(err.output)
             raise StopIteration
         finally:
-            os.remove(driver_file)
+            if not args.keep_driver:
+                os.remove(driver_file)
 
     return
