+++ tools/orac_preproc.py
@@ -323,7 +323,8 @@ def orac_preproc(args):
                     print(err.output)
             raise StopIteration
         finally:
-            os.remove(driver_file)
+            if not args.keep_driver:
+                os.remove(driver_file)
 
     # Return expected output filename
     return outroot
